Coding ~apple(script)
Problem solved, appletree now moves left and right
script of apple done.
static = shared by all instances of a class, so every instance of apple will have the same value. if y is ever changed for one instance, it will simultaneously change for all instance.
this = references the current instance of the C# class in which it is called in this coding list, this reference the apple(script) component instance), not the entire game Object. (this.gameObject)